Patch #5244.
This patch cleans up DATA LIST and PRINT in preparation for
re-implementing REPEATING DATA (bug #12859).  It reduces the code
duplication between DATA LIST and PRINT a great deal, actually
reducing the total code in PSPP by a couple hundred lines.  The result
should be more maintainable as well as easier to read.