Until now, the procedure code has provided a case to the case_source, which has filled in the data values that come from the active file. "Left" data values that don't come from the active file naturally stay the same from case to case, because the procedure code keeps using that same case.
One of the compromises that comes with the new procedure code is that
the active file allocates and provides its own case, which the
procedure code then has to resize to provide room for any other
variables that should go in the case and then fill in the values of
"left" variables.  Then, when we're done with that case, we have to
save the values of "left" variables to copy into the next case read
from the active file.

The caseinit code helps with this.