Complete re-implementation of the EXAMINE command.
The implementation now avoids storing large amounts of data on the
heap.  Instead it uses the casereader functionality and
the casegrouper structure instead of hashes to store group data.
Added documentation for the /ID subcommand which was missing.

Better encapsulated the calculation of boxplot data, extreme
values, histograms, NP plots and percentiles and others. Made
the start of a common base interface for statistic calculations.
Also updated the frequencies code to fit new interfaces.

Added a test for bug #23852