Also warn about missing prerequisites as we encounter them (bug #24445).
It can happen that configure bails out in the middle due to some issue
that PSPP cannot directly control; for example, AC_CHECK_SIZEOF exits
unconditionally if it fails.  If this happens then, before, we would fail
to warn about missing required prerequisites that could well have been
the actual cause of the failure, because we only warn about these at
the end of the run.

Now, we warn about missing prerequisites both at time of detection and
in a summary at the end, to head off this problem.

Thanks to Jim Orr for allowing me to notice the problem.