Replace gsheet-column interface by psppire-axis
Created a new object psppire-axis, which is a thinish
wrapper around src/libpspp/tower.c, and can be used
to maintain the positions and widths of rows/columns
in the data/variable sheets.
This change also deletes the lib/gtksheet/*-column-*
modules, and uses a psppire-axis in their place.
Some functions previously performed by gsheet-column-iface
have been moved to gsheetmodel.
This change doesn't touch lib/gtksheet/*-row-* but it
will be straightforward to replace that too.