perl-module: Make PSPP::Reader::get_next_case() return a list.
PSPP::Reader::get_next_case() was documented to return a list,
but actually it returned a reference to a list.  This commit
changes the behavior to match the documentation.

This also fixes the behavior on reading past the end of the file.
Previously this was documented to return undef but caused a
segfault in practice.  Now it returns an empty list.

Also increase PSPP version number from 0.7.1 to 0.7.2 at John
Darrington's request, to alert users of the previous version of
the Perl interface.