Link against -lws2_32 on Windows.
Gnulib commit df1da811, "Fix link errors on Windows when close
module is used", adds a Makefile variable $(LIB_CLOSE) that expands
to -lws2_32.  This commit adds that variable to the linker command
line.

It might not be necessary to add this to both pspp and psppire,
but it shouldn't hurt.