Make create_iconv() properly distinguish converters by name.
The code in create_iconv() assumed that every pair of different converters
had a different hash value.  This is a bad assumption: eventually, we will
be unlucky, and two different converters will hash to the same value, and
we will get a bad conversion.  So we have to compare (and store) the
names of the codes that each converters converts to and from.

Also, compute the hash value without making an extra copy of fromcode
and tocode.