range-set: Inline some simple functions.
Some of the range-set functions are very simple and worth inlining, so
move the definitions of those functions into range-set.h.  This required
moving the definition of struct range_set and struct range_set_node into
the header.  Some of the functions used internally by those functions had
to be moved too, and renamed as well since for internal use in range-set.c
their names did not respect the namespace.