data-parser: Treat " ," the same as ",".
The data parser (used e.g. by DATA LIST) was unintentionally treating
soft separators (e.g. white space) followed by a hard separator (e.g.
a comma) as an empty field, but in fact it should be treated as just a
single separator.  This fixes it and adds a test.

Thanks to pascal barbedor <pbarbedor@gmail.com> for reporting the problem.