New function parse_value() for parsing a value of specified width.
Occasionally a value of a given width needs to be parsed from syntax.
This commit adds a helper function for doing so and modifies a few pieces
of code to use it.  Probably there are other places where it would be
useful that could not easily be found with "grep".

This commit also renames the range-parser code to value-parser and puts
the new function in there, as a natural generalization.

Suggested by John Darrington.