output: Make table dimension functions work in arbitrary contexts.
Until now the function passed to tab_dim has only been called at the
time that the table is	passed to tab_submit.  But in upcoming commits
the table will be kept around and rendered possibly much later (by the
GUI), so the function to tab_dim has to be sure to only use data that
will exist at that time.  Thus, this commit makes it possible to free
the auxiliary data passed to that function, so that in the few cases
where it needs to keep around extra data, it can do so without leaking
memory.  At the same time, convert the few functions that do need this
feature to use it.