output: Make allocating and freeing drivers independent of configuration.
Until now, the output subsystem has taken responsibility for
allocating, freeing, registering, and unregistering output drivers.
This is usually what we want it to do, but PSPPIRE will soon want to
register an output driver of its own, independent of the main driver
configuration mechanism.  This commit refactors the output driver
subsystem slightly to make this possible.