Make the UTF-8 graphic devices avoid overstriking by default.
The UTF-8 graphic devices introduced in the previous commit use
backspaces and overstriking to indicate bold, but most users probably
won't appreciate that, since it makes it harder to work with the
output files in many text editors.  Thus, disable this behavior.

Thanks to John Darrington for pointing out the issue.