output: Include command names in titles in GUI output.
The command name displayed by the output engine was the current command
name.  The GUI output was never displayed during the actual execution of a
command, so GUI titles never included a command name.

This commit puts the command name into the struct passed into the output
engine, so that it gets saved and displayed.