output: Get rid of histogram_plot_n function.
For some reason we have two functions to plot histograms, one of which is
marked "deprecated" and the other of which is a trivial wrapper.  It's
easy, however, to get rid of one of them and make the caller in charge of
doing what the wrapper does, especially since the wrapper had only a single
caller.  That is what this commit does.