output: Make chart geometry the responsibility of the output driver.
Currently, with libplot, all charts have the same geometry, because
libplot allows us to reasonably set the dimensions of the drawing area
to the same 1000x1000 size.  But cairo will produce strangely scaled
results if we attempt to do the same thing there (distorted font sizes,
etc.).  So make the output drivers responsible for setting up the chart
geometry, so that the output drivers can tell the chart drawing routines
the size of the drawing area.