output: Use Cairo and Pango to draw charts, instead of libplot.
This paves the way for adding support for charts to the PSPPIRE GUI, since
Cairo can render to a window in GTK+ whereas libplot cannot.

I am aware of the following regressions versus the state before this
commit:

1. Alignment of text in charts on the baseline is not properly implemented.
   This makes the labels in pie charts look funny.

2. Chart titles are drawn in the same font size as other chart text.

3. Building without Cairo and Pango is broken.

4. The postscript output driver (as opposed to the cairo output driver in
   PostScript output mode) no longer supports charts.

I plan to fix #1, #2, and #3.  #4 is fixable too, but I don't know whether
it is worth it; I might prefer to delete the postscript output driver
altogether.