Fix missing @clicksequence problem with older Texinfo versions.
The @clicksequence command was introduced in Texinfo 4.12, which a lot of
users do not yet have.  This commit fixes the problem by detecting whether
makeinfo supports @clicksequence and, if not, defining a macro for it as
a substitute.

We also need support for @clicksequence in texinfo.tex.  It is possible to
detect whether @clicksequence is defined using TeX code and, if not,
define a TeX macro for it.  However, I'm not sure how to do that.  So this
commit instead adds a known-good texinfo.tex to the source tree, which
should have the same effect.