Don't pad variable labels written to system files.
Previously, the code to write variable labels to system files accidentally
padded out the variable labels on the right with spaces to a multiple of 4
bytes in length.  With this commit, variable labels will no longer be
padded.

This fixes a genuine bug for variable labels longer than 252 bytes, which
previously were padded out to 256 bytes.  Variable labels are limited to
a maximum of 255 bytes, so PSPP refused to read such files.

Reported-by: Robert Westlund <rob@kinesissurvey.com>