Added a new target "programs" to the makefile.
When a source file is changed, the "all" target (correctly) rebuilds
the po/pspp.pot file and remerges the PO files for all supported translations,
and then builds .gmo files for each of them.   Whilst this is correct behaviour,
it will become an unmitigated nuisance for developers as more and more
translations are added.

This change adds a new target called "programs" which builds the binaries, but
not the ancilliary targets. This target is more convienient for working on
changes to the source code, when translations are not of interest.