Fix numerous warnings about an unknown attribute rpl_malloc on mingw32.
On mingw32, Gnulib replaces malloc by defining it as a macro that expands
to rpl_malloc.  Since MALLOC_LIKE used malloc as an identifier, even though
it didn't refer to the malloc function per se, this instance of malloc
also expanded to rpl_malloc, confusing GCC.

GCC also accepts __malloc__ here, so this commit uses that spelling
instead, which avoids the warning.