Fix get-data-txt-importcases.sh test on Mingw.
On Mingw, ftell() and ftello() are unreliable when used on text files with
plain LF line endings:
    http://article.gmane.org/gmane.comp.gnu.m4.bugs/2224

This made the get-data-txt-importcases.sh test fail, because
dfm_get_percent_read() that it depends upon to determine the percentage of
the file that has been read uses ftello().  The solution is to always open
the text file as a binary file.  According to my audit of the ways that
the text file is read, this should not cause other problems.  In
particular, ds_read_line() handles both LF and CR-LF line endings.  But it
cannot handle negative effects on Unix platforms anyway, since they do not
distinguish text and binary modes.