message: Fix corner case in msg_emit().
In most cases, msg_emit() initialized the "where" member of the message
passed in.  However, if msg_init() had not yet been called, it did not do
this.  This fixes that corner case.

In addition, one caller of msg_emit() did not initialize these members.  It
is not necessary for both msg_emit() and its callers to do so, but since
most callers did so, this commit simply makes the remaining caller do so
as well, for consistency.