pspp: Fix interactive usability for multi-line commands.
Until now, the terminal UI has only flushed output just before reading
the first line of a command.  This generally worked OK because typically
only the final command produced output, not intermediate lines.  But now
every line of a command potentially produces output (e.g. the command name
or the command syntax).  With the ASCII driver that makes the interface
difficult to use because it typically opens and holds open a pipe to an
interactive program such as "more", which in turn turns off terminal echo
etc.

Flushing output before reading each command avoids this problem and makes
the terminal UI more usable.