configure.ac: Restore use of AM_GNU_GETTEXT.
AM_GNU_GETTEXT is useful for its sophisticated support for finding the
libraries needed to use gettext, but it also calls AM_PO_SUBDIRS, which
we do not want because PSPP has its own "po" directory makefile support.

This commit reintroduces AM_GNU_GETTEXT but disables AM_PO_SUBDIRS by
declaring that we have already called it using AC_PROVIDE.  It also
adds a dummy Makefile.in to the "po" directory and an ABOUT-NLS file
because Automake requires both.